refactor: rename lib folder to utils folder
