+++ src/utils/aspectRatioHelper.ts
